### 2026-01-16 #744 @mccroweyclinton-EPA
request for R package RAQSAPI submission
Submitting Author Name: Clinton Mccrowey
Submitting Author Github Handle: @MCCROWEYCLINTON-EPA
Repository:  https://github.com/USEPA/RAQSAPI
Version submitted: v2.0.5
Submission type: Standard

Version accepted: TBD

---

```
A R extension to retrieve EPA Air Quality System Data via the AQS Data Mart API
```

RAQSAPI is a R package that was designed specifically as a simple interface to the US Environmental Protection Agency Air Quality System Data Mart API which is a data store of ambient air pollution data collected by EPA, state, local, and tribal air pollution control agencies from over thousands of monitors. AQS also contains meteorological data, descriptive information about each monitoring station (including its geographic location and its operator), and data quality assurance/quality control information.

Environmental scientists, data scientist, Federal, State, Local, Tribal Air monitoring agencies. Community groups. Citizens concerned about the quality of air in the US.

[aqsr](https://github.com/kpkeller/aqsr), RAQSAPI is still under active development and will be updated reflecting future changes to the AQS Data MART API. RAQSAPI is developed by EPA staff. RAQSAPI allows for multi year data pulls. Also RAQSAPI is the only package that I am aware of that is available from CRAN.

YES

https://github.com/ropensci/software-review/issues/712

Currently I have pkgcheck to only check for issues in the main branch most of these issues are old, some have been resolved in other branches.

 ✖️ Package coverage failed
 ✖️ R CMD check found 1 error

I believe both of these issues are related to issues that I am currently having with implementing mock test with R package:httptest2. Several members of the Ropensci team is currently helping me to work out these issues.

✖️ does not have a 'contributing' file.
currently pkgcheck only checks for issues in the main branch. the branch "test_httptest2" contains the current version of the contributing.md file and will be merged into main in the future. If required now I can commit it to main.
