### 2026-01-20 19ea57b
add CODE_OF_CODUCT.md and contributing.md to main branch, minor edits to .gitignore
