### 2026-01-20 ff108eb
remove importing testthat, httptest2 in unit test files, remove useless code in redact.R - as instructed by the creator of httptest2 in #29
