+++ tests/testthat/test-byMA.R
@@ -1,7 +1,4 @@
 #' @importFrom magrittr `%>%`()
-#' @import testthat
-#' @import httptest2
-
 if (file.exists("local.R"))
 {
   source("helper.R")
