+++ tests/testthat/test-bystate.R
@@ -1,8 +1,4 @@
 #' @importFrom magrittr `%>%`()
-#' @import testthat
-#' @import httptest2
-
-
 if (file.exists("local.R"))
 {
   source("helper.R")
