### 2026-01-20 9635429
remove code that redacts credentials from tests/testthat/setup.R
