+++ tests/testthat/setup.R
@@ -9,15 +9,3 @@ httptest2::.mockPaths(c("./tests/testthat/",
   "./tests/testthat/bysite",
   "./tests/testthat/bystate",
   "./tests/testthat/listfunctions"))
-
-# for contexts where the package needs to be fooled
-# (CRAN, forks)
-# this is ok because the package will used recorded responses
-# so no need for a real secret
-if (!nzchar(Sys.getenv("RAQSAPIKEY"))) {
-  Sys.setenv(RAQSAPIKEY = "<redacted>")
-}
-
-if (!nzchar(Sys.getenv("RAQSAPIUSERNAME"))) {
-  Sys.setenv(RAQSAPIUSERNAME = "<redacted>")
-}
