+++ vignettes/TimelinessofAQSData.Rmd
@@ -18,7 +18,7 @@ csl: acs-nano.csl
 
 # About the timeliness of AQS Data
 
-EPA's AQS Datamart API, the service that RAQSAPI retrieves data from, does not
+EPA's AQS DataMart API, the service that RAQSAPI retrieves data from, does not
 host real time (collected now/today) data. If real time data is needed, please
 use the AirNow API and direct all questions toward real time data there. RAQSAPI
 does not work with AirNow and cannot retrieve real time data. For more details
