+++ man/aqs_sampledata_by_site.Rd
@@ -41,7 +41,7 @@ list of available county codes for each state.}
 \item{sitenum}{a R character object which represents the 4 digit site number
 (with leading zeros) within the county and state being
 requested. @seealso \code{\link[=aqs_sites_by_county]{aqs_sites_by_county()}} for the list of
-availables site numbers for a given county and state.}
+available site numbers for a given county and state.}
 
 \item{duration}{an optional R character string that represents the
 parameter duration code that limits returned data
