+++ CITATION.cff
@@ -31,7 +31,7 @@ keywords:
 - United State Environmental Protection Agency
 - Environmental-data-retrieval
 - Air-Quality-System
-- AQS-Datamart
+- AQS-DataMart
 - Air-monitoring
 - Environmental-monitoring
 - airmonitoring
