+++ tests/testthat/test-byMA.R
@@ -13,8 +13,8 @@ if (file.exists("local.R"))
 RAQSAPI::aqs_credentials(username = datamartAPI_user, key = datamartAPI_key)
 
 
-with_mock_api({
-#with_mock_dir("byMA", {
+#with_mock_api({
+with_mock_dir("byMA", {
   test_that(
     "byMA functions", {
 
