+++ tests/testthat/test-AQS_DATAMART_API-S3-class.R
@@ -1,6 +1,6 @@
 #' @importFrom magrittr `%>%`()
 #' @importFrom tibble tibble
-#' @importFrom lubridate now
+#' @importFrom lubridate now leap_year
 test_that(
   "AQS_DATAMART_API S3 class", {
     .Data <- tibble::tibble(a = 1:10, b = 11:20, c = 21:30)
@@ -23,6 +23,7 @@ test_that(
 #' @importFrom lubridate mdy_hms now year
 #' @import from glue glue
 #' @import testthat
+#' @importFrom lubridate now year mdy_hms
 test_that(
   "test AQS_DATAMART_APIv2_validator", {
 
@@ -30,13 +31,13 @@ test_that(
       year()
     fakeData <- tibble(
       datetime = seq.POSIXt(
-        from = mdy_hms(glue("01-01-{now() %>% year()} 00:00:00")),
-        to = mdy_hms(glue("12-31-{now() %>% year()} 23:59:59")),
+        from = lubridate::mdy_hms(glue("01-01-{lubridate::now() %>% year()} 00:00:00")),
+        to = lubridate::mdy_hms(glue("12-31-{lubridate::now() %>% year()} 23:59:59")),
         by = "hour"
       ),
       sample = rnorm(
         n = ifelse(
-          leap_year(year),
+          lubridate::leap_year(year),
           87840, 8760
         ),
         mean = 100, sd = 50
