### 2026-04-03 61c618d
addresses #29, code cleanup and linting
