+++ DESCRIPTION
@@ -85,8 +85,7 @@ Suggests: cli,
     roxygen2,
     spelling,
     testthat (>= 3.0.0),
-    usethis,
-    withr
+    usethis
 SystemRequirements: package manual and Vignettes require pandoc
                     (>= 1.14) http://pandoc.org)
 RoxygenNote: 7.3.3
@@ -116,8 +115,9 @@ Collate:
 Copyright: United States Environmental Protection Agency
 RdMacros: lifecycle
 Roxygen: list(markdown = TRUE)
-Config/testthat/edition: 3
 X-schema.org-keywords: United State Environmental Protection Agency,
                        Environmental-data-retrieval, Air-Quality-System,
                        AQS-DataMart, Air-monitoring, Environmental-monitoring
+Config/testthat/edition: 3
 Config/testthat/load-all: list(export_all = FALSE, helpers = FALSE)
+Config/testthat/parallel: false

[…]
