### 2026-04-03 3be9e11
rebuild README
