+++ README.md
@@ -1,0 +1,600 @@
+Introduction to the RAQSAPI package
+================
+Clinton Mccrowey, physical scientist - US EPA
+
+- [EPA Disclaimer](#epa-disclaimer)
+- [Introduction](#introduction)
+- [About the timeliness of AQS Data](#about-the-timeliness-of-aqs-data)
+- [Installing RAQSAPI](#installing-raqsapi)
+  - [Option 1: Installing the stable version from CRAN (preferred
+    method)](#option-1-installing-the-stable-version-from-cran-preferred-method)
+  - [Option 2: Installing the development version of RAQSAPI
+    (unstable)](#option-2-installing-the-development-version-of-raqsapi-unstable)
+- [Using The RAQSAPI library](#using-the-raqsapi-library)
+  - [Load RAQSAPI](#load-raqsapi)
+  - [Sign up and setting up user credentials with the RAQSAPI
+    library](#sign-up-and-setting-up-user-credentials-with-the-raqsapi-library)
+  - [(suggested) Use the `keyring` package to manage
+    credentials](#suggested-use-the-keyring-package-to-manage-credentials)
+  - [Data Mart aggregate functions](#data-mart-aggregate-functions)
+    - [See the RAQSAPI vignette for more
+      details](#see-the-raqsapi-vignette-for-more-details)
+  - [pyaqsapi - a port of RAQSAPI to the python 3 programming
+    environment](#pyaqsapi---a-port-of-raqsapi-to-the-python-3-programming-environment)
+- [Acknowledgements](#acknowledgements)
+- [References](#references)
+
+<img src="./vignettes/figures/RAQSAPIhexsticker.png" alt="RAQSAPI hexsticker" align="right" width="150" height="150">
+
+<!-- badges: start -->
+
+[![Project Status: Active – The project has reached a stable, usable
+state  
+and is being actively
+developed.](https://www.repostatus.org/badges/latest/active.svg)](https://www.repostatus.org/#active)
+[![R-CMD-check](https://github.com/USEPA/RAQSAPI/workflows/R-CMD-check/badge.svg)](https://github.com/USEPA/RAQSAPI/actions)
+[![CRAN
+status](https://www.r-pkg.org/badges/version/RAQSAPI)](https://CRAN.R-project.org/package=RAQSAPI)
+[![metacran
+downloads](https://cranlogs.r-pkg.org/badges/RAQSAPI)](https://cran.r-project.org/package=RAQSAPI)
+[![lifecycle](https://img.shields.io/badge/lifecycle-maturing-blue.svg)](https://lifecycle.r-lib.org/articles/stages.html)
+[![license](https://img.shields.io/badge/license-MIT-green)](https://choosealicense.com/licenses/mit/)
+[![Last-changedate](https://img.shields.io/badge/last%20change-%202026--04--03-yellowgreen.svg)](/commits/master)
+[![pkgcheck](https://github.com/USEpa/RAQSAPI/workflows/pkgcheck/badge.svg)](https://github.com/USEpa/%20/RAQSAPI/.github/workflows/pkgcheck.yaml)
+<!-- badges: end -->
+
+# EPA Disclaimer
+
+> \[!NOTE\]
+>
+> This software/application was developed by the U.S. Environmental
+> Protection Agency (USEPA). No warranty expressed or implied is made
+> regarding the accuracy or utility of the system, nor shall the act of
+> distribution constitute any such warranty. The USEPA has relinquished
+> control of the information and no longer has responsibility to protect
+> the integrity, confidentiality or availability of the information. Any
+> reference to specific commercial products, processes, or services by
+> service mark, trademark, manufacturer, or otherwise, does not
[…]
