### 2026-04-06 bc17770
trival edits to .github/workflows/R-CMD-check.yaml, use dummy variable in mocked unit tests
