+++ tests/testthat/test-bypqao.R
@@ -10,6 +10,14 @@ if (file.exists("local.R"))
     datamartAPI_user <- Sys.getenv("RAQSAPIUSERNAME", names = TRUE)
     datamartAPI_key <- Sys.getenv("RAQSAPIKEY", names = TRUE)
   }
+
+# If credentials are not available (e.g., during check_built), set dummy values for mocked tests
+if (is.na(datamartAPI_user) || datamartAPI_user == "" || is.na(datamartAPI_key) || datamartAPI_key == "")
+{
+  datamartAPI_user <- "test@example.com"
+  datamartAPI_key <- "testkey"
+}
+
 RAQSAPI::aqs_credentials(username = datamartAPI_user, key = datamartAPI_key)
 
 
