+++ .github/workflows/deploy_to_gh-pages.yaml
@@ -1,40 +1,0 @@
-name: deploy to github pages
-description: "prepare RAQSAPI vignettes for publishing on github pages"
-on:
-  workflow_run:
-    workflows: ["R-CMD-check.yaml"]
-    branches: [main]
-    types:
-      - completed
-  workflow_dispatch:
-
-jobs:
-  install_Pandoc:
-    runs-on: ubuntu-latest
-    steps:
-      - name: Install Pandoc
-        run: sudo apt-get update && sudo apt-get install -y pandoc
-
-  build:
-    runs-on: ubuntu-latest
-    needs: install_Pandoc
-    if: ${{ github.event.workflow_run.conclusion == 'success' }}
-    steps:
-      - name: Build static files
-        run: |
-          R CMD build . --md5 --compact-vignettes
-
-      - name: Upload static files as artifact
-        uses: actions/upload-pages-artifact@v3
-        with:
-          path: vignettes/*.md
-
-  deploy:
-    environment:
-      name: github-pages
-    runs-on: ubuntu-latest
-    needs: build
-    steps:
-      - name: Deploy to GitHub Pages
-        id: deployment
-        uses: actions/deploy-pages@v4
