### 2026-04-27 9eaa758
replace remotes package with pak, update authors with ORCID and RORs, config/build/clean-inst/doc:true, fix issues with github workflow files, fully qualify references to internal roxygen2 references, add trival unit tests
