+++ vignettes/RAQSAPIPurpose.R
@@ -1,5 +1,0 @@
-## ----setup, include = FALSE------------------------------------------------------------------------------------------------
-knitr::opts_chunk$set(
-  collapse = TRUE,
-  comment = "#>"
-)
