+++ vignettes/pyaqsapi.R
@@ -1,1 +1,0 @@
-## ----setup, include=FALSE-----------------------------------------------------
