+++ R/setupfunctions.R
@@ -4,7 +4,7 @@
 #'                 needs to be called once and only once every time this library
 #'                 is re-loaded. Users must have a valid username and key which
 #'                 can be obtained through the use of the aqs_sign_up function,
-#'                 @seealso [aqs_sign_up()] to sign up for AQS data mart
+#'                 @seealso [RAQSAPI::aqs_sign_up()] to sign up for AQS data mart
 #'                 credentials.
 #' @importFrom rlang is_character
 #' @param username a R character object which represents the email account that
@@ -54,7 +54,7 @@ aqs_credentials <- function(username = NA_character_, key = NA_character_)
 #'                 the account specified. Follow the instructions
 #'                 in the verification e-mail before proceeding to use any other
 #'                 functionality of the AQS API. Register your credential
-#'                 with [aqs_credentials()] before using the
+#'                 with [RAQSAPI::aqs_credentials()] before using the
 #'                 other functions in this library.
 #' @note The '@' character needs to be escaped with the '/' character.
 #' @importFrom glue glue
