+++ vignettes/AQSAPI.bib
@@ -123,13 +123,13 @@
     doi = {10.32614/CRAN.package.lintr},
   }
 
-@Manual{package:remotes,
-    title = {lintr: A 'Linter' for R Code},
-    author = {Jim Hester and Florent Angly and Russ Hyde and Michael Chirico and Kun Ren and Alexander Rosenstock and Indrajeet Patil},
-    year = {2025},
-    note = {R package version 3.2.0},
-    url = {https://CRAN.R-project.org/package=lintr},
-    doi = {10.32614/CRAN.package.lintr},
+  @Manual{package:pak,
+    title = {pak: Another Approach to Package Installation},
+    author = {Gabor Csardi and Jim Hester},
+    year = {2026},
+    note = {R package version 0.9.5},
+    url = {https://CRAN.R-project.org/package=pak},
+    doi = {10.32614/CRAN.package.pak},
   }
 
  @Misc{py3package:pyaqsapi,

[…]
