+++ inst/httptest2/redact.R
@@ -4,19 +4,35 @@ function(resp)
   user <- getOption("aqs_username")
 
   if (!is.null(key) && nzchar(as.character(key))) {
-    resp <- httptest2::gsub_response(resp, as.character(key), "redacted", fixed = TRUE)
+    resp <- httptest2::gsub_response(resp,
+                                     as.character(key),
+                                     "redacted", fixed = TRUE)
   }
   if (!is.null(user) && nzchar(as.character(user))) {
-    resp <- httptest2::gsub_response(resp, as.character(user), "redacted", fixed = TRUE)
+    resp <- httptest2::gsub_response(resp,
+                                     as.character(user),
+                                     "redacted", fixed = TRUE)
   }
 
   # shorten all mock paths by abbreviating the base URL
-  resp <- httptest2::gsub_response(resp, "https://aqs.epa.gov/data/api", "aqs", fixed = TRUE)
+  resp <- httptest2::gsub_response(resp,
+                                   "https://aqs.epa.gov/data/api",
+                                   "aqs",
+                                   fixed = TRUE)
   # abbreviate long function, changes path in mocks
-  resp <- httptest2::gsub_response(resp, "qaAnnualPerformanceEvaluations", "QAape", fixed = TRUE)
+  resp <- httptest2::gsub_response(resp,
+                                   "qaAnnualPerformanceEvaluations",
+                                   "QAape",
+                                   fixed = TRUE)
   # abbreviate long function, causes path in mocks
-  resp <- httptest2::gsub_response(resp, "transactionsQaAnnualPerformanceEvaluations", "tQAape", fixed = TRUE)
-  resp <- httptest2::gsub_response(resp, "quarterlyData", "QD", fixed = TRUE)
+  resp <- httptest2::gsub_response(resp,
+                                   "transactionsQaAnnualPerformanceEvaluations",
+                                   "tQAape",
+                                   fixed = TRUE)
+  resp <- httptest2::gsub_response(resp,
+                                   "quarterlyData",
+                                   "QD",
+                                   fixed = TRUE)
 
   return(resp)
 }
