+++ .github/workflows/R-CMD-check.yaml
@@ -6,7 +6,7 @@ on:
   pull_request:
   workflow_dispatch:
 
-name: R-CMD-check.yaml
+name: R-CMD-check
 
 permissions: read-all
 
