### 2026-04-27 0f04ec8
update inst/WORDLIST for spelling checks, fix .github/workflows/pkgcheck.yaml workflow dependency
