### 2026-04-27 #744 @mccroweyclinton-EPA
@ldecicco-USGS or @mpadge, Please re-run the pkgcheck bot, I think I may have fixed all issues except the don't use dontrun checks that pkgcheck throws, Since functions in this package make API calls to a server and require user credentials to run any function, I have no choice but to use dontrun on almost all of the examples included in the roxygen comments. I have merged the fixed code into the main branch.
