### 2026-04-28 1baf087
update pkgcheck workflow to checkout code and retrieve PR numbers, spell_check_test in tests/spelling.R now points to the correct path
