+++ tests/spelling.R
@@ -1,2 +1,8 @@
-if (requireNamespace("spelling", quietly = TRUE)) spelling::spell_check_test(vignettes = TRUE,
-                                                                             error = FALSE, skip_on_cran = TRUE)
+if (requireNamespace("spelling", quietly = TRUE))
+  {
+    spelling::spell_check_test(pkg = "../",
+                               vignettes = TRUE,
+                               error = FALSE,
+                               lang = "en-US",
+                               skip_on_cran = FALSE)
+  }
