### 2026-04-28 cfb1081
fix lints and changes to .lintr config file
