+++ R/bycounty.R
@@ -26,18 +26,18 @@
 #'                                  )
 #'          }
 #' @export
-aqs_monitors_by_county <- function(parameter, bdate, edate, stateFIPS, countycode, cbdate = NA_Date_,
-                                   cedate = NA_Date_, return_header = FALSE)
+aqs_monitors_by_county <- function(parameter, bdate, edate, stateFIPS, countycode, cbdate = lubridate::NA_Date_,
+                                   cedate = lubridate::NA_Date_, return_header = FALSE)
 {
   checkaqsparams(parameter, bdate, edate, stateFIPS, countycode, cbdate, cedate, return_header)
 
   # aqs_monitors_by_* functions don't call aqsmultiyearparams() since the monitors API call accepts multiple years of data
   # on the server, purrr::pmap is used so that the output is consistent with other RAQSAPI functions.
-  params <- tibble(
+  params <- tibble::tibble(
     parameter = parameter, bdate = bdate, edate = edate, stateFIPS = stateFIPS,
     countycode = countycode, service = "monitors", cbdate = cbdate, cedate = cedate
   ) %>%
-    select(where(~!all(is.na(.x))))
+    dplyr::select(tidyselect::where(~!all(is.na(.x))))
 
   monitors <- purrr::pmap(.l = params, .f = aqs_services_by_county)
   if (!return_header)
@@ -104,8 +104,8 @@ aqs_sampledata_by_county <- function(parameter,
   stateFIPS,
   countycode,
   duration = NA_character_,
-  cbdate = NA_Date_,
-  cedate = NA_Date_,
+  cbdate = lubridate::NA_Date_,
+  cedate = lubridate::NA_Date_,
   return_header = FALSE
 )
 {
@@ -171,7 +171,7 @@ aqs_sampledata_by_county <- function(parameter,
 #'          }
 #' @export
 aqs_annualsummary_by_county <- function(parameter, bdate, edate, stateFIPS, countycode,
-                                        cbdate = NA_Date_, cedate = NA_Date_, return_header = FALSE)
+                                        cbdate = lubridate::NA_Date_, cedate = lubridate::NA_Date_, return_header = FALSE)
 {
   checkaqsparams(parameter, bdate, edate, stateFIPS, countycode, cbdate, cedate, return_header)
 
@@ -233,8 +233,8 @@ aqs_annualsummary_by_county <- function(parameter, bdate, edate, stateFIPS, coun
 #'                                    )
 #'            }
 #' @export
-aqs_qa_blanks_by_county <- function(parameter, bdate, edate, stateFIPS, countycode, cbdate = NA_Date_,
-                                    cedate = NA_Date_, return_header = FALSE)
+aqs_qa_blanks_by_county <- function(parameter, bdate, edate, stateFIPS, countycode, cbdate = lubridate::NA_Date_,
+                                    cedate = lubridate::NA_Date_, return_header = FALSE)
 {
   checkaqsparams(parameter, bdate, edate, stateFIPS, countycode, cbdate, cedate, return_header)
 
@@ -296,7 +296,7 @@ aqs_qa_blanks_by_county <- function(parameter, bdate, edate, stateFIPS, countyco
 #'          }
 #' @export
 aqs_dailysummary_by_county <- function(parameter, bdate, edate, stateFIPS, countycode,
-                                       cbdate = NA_Date_, cedate = NA_Date_, return_header = FALSE)
+                                       cbdate = lubridate::NA_Date_, cedate = lubridate::NA_Date_, return_header = FALSE)
 {
   checkaqsparams(parameter, bdate, edate, stateFIPS, countycode, cbdate, cedate, return_header)
 
[…]
