+++ R/bystate.R
@@ -27,18 +27,18 @@
 #'           }
 #'
 #' @export
-aqs_monitors_by_state <- function(parameter, bdate, edate, stateFIPS, cbdate = NA_Date_,
-                                  cedate = NA_Date_, return_header = FALSE)
+aqs_monitors_by_state <- function(parameter, bdate, edate, stateFIPS, cbdate = lubridate::NA_Date_,
+                                  cedate = lubridate::NA_Date_, return_header = FALSE)
 {
   checkaqsparams(parameter, bdate, edate, stateFIPS, cbdate, cedate, return_header)
 
   # aqs_monitors_by_* functions don't call aqsmultiyearparams() since the monitors API call accepts multiple years of data
   # on the server, purrr::pmap is used so that the output is consistent with other RAQSAPI functions.
-  params <- tibble(
+  params <- tibble::tibble(
     parameter = parameter, bdate = bdate, edate = edate, stateFIPS = stateFIPS,
     service = "monitors", cbdate = cbdate, cedate = cedate
   ) %>%
-    select(where(~!all(is.na(.x))))
+    dplyr::select(tidyselect::where(~!all(is.na(.x))))
 
   monitors <- purrr::pmap(.l = params, .f = aqs_services_by_state)
   if (!return_header)
@@ -101,7 +101,7 @@ aqs_monitors_by_state <- function(parameter, bdate, edate, stateFIPS, cbdate = N
 #'                    }
 #' @export
 aqs_sampledata_by_state <- function(parameter, bdate, edate, stateFIPS, duration = NA_character_,
-                                    cbdate = NA_Date_, cedate = NA_Date_, return_header = FALSE)
+                                    cbdate = lubridate::NA_Date_, cedate = lubridate::NA_Date_, return_header = FALSE)
 {
   checkaqsparams(parameter, bdate, edate, stateFIPS, duration, cbdate, cedate, return_header)
 
@@ -165,8 +165,8 @@ aqs_sampledata_by_state <- function(parameter, bdate, edate, stateFIPS, duration
 #'                                               )
 #'                    }
 #' @export
-aqs_annualsummary_by_state <- function(parameter, bdate, edate, stateFIPS, cbdate = NA_Date_,
-                                       cedate = NA_Date_, return_header = FALSE)
+aqs_annualsummary_by_state <- function(parameter, bdate, edate, stateFIPS, cbdate = lubridate::NA_Date_,
+                                       cedate = lubridate::NA_Date_, return_header = FALSE)
 {
   checkaqsparams(parameter, bdate, edate, stateFIPS, cbdate, cedate, return_header)
 
@@ -227,8 +227,8 @@ aqs_annualsummary_by_state <- function(parameter, bdate, edate, stateFIPS, cbdat
 #'                                           )
 #'                    }
 #' @export
-aqs_qa_blanks_by_state <- function(parameter, bdate, edate, stateFIPS, cbdate = NA_Date_,
-                                   cedate = NA_Date_, return_header = FALSE)
+aqs_qa_blanks_by_state <- function(parameter, bdate, edate, stateFIPS, cbdate = lubridate::NA_Date_,
+                                   cedate = lubridate::NA_Date_, return_header = FALSE)
 {
   checkaqsparams(parameter, bdate, edate, stateFIPS, cbdate, cedate, return_header)
 
@@ -289,8 +289,8 @@ aqs_qa_blanks_by_state <- function(parameter, bdate, edate, stateFIPS, cbdate = 
 #'                                              )
 #'                    }
 #' @export
-aqs_dailysummary_by_state <- function(parameter, bdate, edate, stateFIPS, cbdate = NA_Date_,
[…]
