### 2026-04-29 #744 @mccroweyclinton-EPA
Please let me know when the bot is able to perform its checks, until then @mpadge, I remember you suggesting that I should clean up the checkaqsparams function in R/AQSAPI_helperfunctions.R to improve the cyclocomplexity but I don't remember where you posted that at. Can you help me with implementing a less complex way of validating user inputs.
