### 2026-05-03 #744 @maurolepore
I see @ldecicco-USGS is actively handling it.

Also I see the bot is struggling to run pkgcheck (FYI @mpadge). If necessary [here're the docs to run it locally](https://docs.ropensci.org/pkgcheck/).

Looking forward to seeing this move through review.

Thanks for sharing your work with the rOpenSci community.
