### 2026-05-04 #744 @ldecicco-USGS
I pulled down the repository and tried running R check via devtools:

```r
==> devtools::check(args = c('--manual','=','TRUE','--as-cran'), build_args = c('--md5','--log','--compact-vignettes'))

══ Building ════════════════════════════════════════════════════════════
Setting env vars:
• CFLAGS    : -Wall -pedantic -fdiagnostics-color=always
• CXXFLAGS  : -Wall -pedantic -fdiagnostics-color=always
• CXX11FLAGS: -Wall -pedantic -fdiagnostics-color=always
• CXX14FLAGS: -Wall -pedantic -fdiagnostics-color=always
• CXX17FLAGS: -Wall -pedantic -fdiagnostics-color=always
• CXX20FLAGS: -Wall -pedantic -fdiagnostics-color=always
── R CMD build ─────────────────────────────────────────────────────────
✔  checking for file 'C:\Users\ldecicco\Documents\RAQSAPI/DESCRIPTION' (578ms)
─  preparing 'RAQSAPI': (2s)
✔  checking DESCRIPTION meta-information (810ms)
─  installing the package (it is needed to process help pages) (767ms)
✔  creating vignettes (38.6s)
─  checking for LF line-endings in source and make files and shell scripts (5.8s)
─  checking for empty or unneeded directories (648ms)
─  adding MD5 file (6.3s)
─  building 'RAQSAPI_2.0.6.tar.gz' (5.7s)

══ Checking ════════════════════════════════════════════════════════════
Setting env vars:
• _R_CHECK_CRAN_INCOMING_REMOTE_               : FALSE
• _R_CHECK_CRAN_INCOMING_                      : FALSE
• _R_CHECK_FORCE_SUGGESTS_                     : FALSE
• _R_CHECK_PACKAGES_USED_IGNORE_UNUSED_IMPORTS_: FALSE
• NOT_CRAN                                     : true
Error in if (file.info(path)$isdir) { : the condition has length > 1
Calls: <Anonymous> -> check_built -> <Anonymous> -> force -> <Anonymous>
Execution halted

Exited with status 1.

```
I'm getting that in 4.5 and 4.6, and I'm really stuck on where the error is coming from.

It's building fine, when I run `devtools::test()` everything passes. Poking around all the functions everything looks fine.

I'm going to go ahead and push on to the next steps (looking for reviewers), and while I do that we can keep troubleshooting why this package won't run a simple package check on my machine (yet it works just fine in GitHub Actions). Maybe whatever is giving me a local headache is also causing the ropensci bot to have issues.
