### 2026-05-12 9e33690
use air formatter, fix lint with lintr, minor changes to .editorconfig, DESCRIPTION and .github/workflows.pkgcheck.yaml
