+++ tests/testthat/test-helperfunctions.R
@@ -1,20 +1,17 @@
 #' @importFrom magrittr `%>%`()
 #' @importFrom stringr str_detect
-if (file.exists("local.R"))
-{
+if (file.exists("local.R")) {
   source("helper.R")
   AQScredentials <- RAQSAPItestsetup_helper()
   datamartAPI_user <- AQScredentials$datamartAPI_user
   datamartAPI_key <- AQScredentials$datamartAPI_key
-} else
-  {
-    datamartAPI_user <- Sys.getenv("RAQSAPIUSERNAME", names = TRUE)
-    datamartAPI_key <- Sys.getenv("RAQSAPIKEY", names = TRUE)
-  }
+} else {
+  datamartAPI_user <- Sys.getenv("RAQSAPIUSERNAME", names = TRUE)
+  datamartAPI_key <- Sys.getenv("RAQSAPIKEY", names = TRUE)
+}
 
 # If credentials are not available (e.g., during check_built), set dummy values for mocked tests
-if (is.na(datamartAPI_user) || datamartAPI_user == "" || is.na(datamartAPI_key) || datamartAPI_key == "")
-{
+if (is.na(datamartAPI_user) || datamartAPI_user == "" || is.na(datamartAPI_key) || datamartAPI_key == "") {
   datamartAPI_user <- "test@example.com"
   datamartAPI_key <- "testkey"
 }
@@ -22,64 +19,62 @@ if (is.na(datamartAPI_user) || datamartAPI_user == "" || is.na(datamartAPI_key) 
 RAQSAPI::aqs_credentials(username = datamartAPI_user, key = datamartAPI_key)
 
 
-test_that(
-  "helperfunctions (checkaqsparams()) functions", {
-    server <- "AQSDatamartAPI"
+test_that("helperfunctions (checkaqsparams()) functions", {
+  server <- "AQSDatamartAPI"
 
-
-    RAQSAPI:::checkaqsparams(service = 99999) %>%
-      expect_error()
-    RAQSAPI:::checkaqsparams(service = "notanactualservice") %>%
-      expect_error()
-    RAQSAPI:::checkaqsparams(parameter = "abcdefg") %>%
-      expect_error()
-    RAQSAPI:::checkaqsparams(bdate = "notadate") %>%
-      expect_error()
-    RAQSAPI:::checkaqsparams(edate = "notadate") %>%
-      expect_error()
-    RAQSAPI:::checkaqsparams(cbdate = "notadate") %>%
-      expect_error()
-    RAQSAPI:::checkaqsparams(cedate = "notadate") %>%
-      expect_error()
-    RAQSAPI:::checkaqsparams(stateFIPS = "90210") %>%
-      expect_error()
-    RAQSAPI:::checkaqsparams(countycode = "341478") %>%
-      expect_error()
-    RAQSAPI:::checkaqsparams(sitenum = "1") %>%
-      expect_error()
-    RAQSAPI:::checkaqsparams(cbsa_code = "something") %>%
-      expect_error()
-    RAQSAPI:::checkaqsparams(pqao_code = "99999") %>%
-      expect_error()
-    RAQSAPI:::checkaqsparams(minlat = "3245253") %>%
-      expect_error()
-    RAQSAPI:::checkaqsparams(maxlat = "647352") %>%
-      expect_error()
-    RAQSAPI:::checkaqsparams(minlon = 45425252) %>%
-      expect_error()
-    RAQSAPI:::checkaqsparams(maxlon = 463753415) %>%
-      expect_error()
-    RAQSAPI:::checkaqsparams(MA_code = "MA") %>%
-      expect_error()
-    RAQSAPI:::checkaqsparams(return_header = 1) %>%
[…]
