+++ R/setupfunctions.R
@@ -22,21 +22,18 @@
 #'                               key = 'MyKey')
 #'          }
 #' @export
-aqs_credentials <- function(username = NA_character_, key = NA_character_)
-{
+aqs_credentials <- function(username = NA_character_, key = NA_character_) {
   # nocov The code simply stores the credentials as a R option. Since the Data Mart server only issues a 'key' and not a
   # 'password' we don't need to worry about securing the credentials with complicated code such as involving salt and hashes
   # and etc.
-  if (rlang::is_character(username) && rlang::is_character(key) && !is.na(username) && !is.na(key))
-  {
+  if (rlang::is_character(username) && rlang::is_character(key) && !is.na(username) && !is.na(key)) {
     options(aqs_username = username)
     options(aqs_key = key)
-  } else
-    {
-      warning("Please enter a valid username and key  \n")
-    }
+  } else {
+    warning("Please enter a valid username and key  \n")
+  }
   return(invisible())
-}  #no cov end
+} #no cov end
 
 
 #' @title aqs_sign_up
@@ -69,8 +66,7 @@ aqs_credentials <- function(username = NA_character_, key = NA_character_)
 #' @return NULL, This functions is called for its side effect and
 #'         does not return meaningful data.
 #' @export
-aqs_sign_up <- function(email)
-{
+aqs_sign_up <- function(email) {
   # nocov start We do not want aqs_sign_up registering new users as part of the unit testing procedures.
 
   # user_agent <- glue('User:{email} via RAQSAPI-{packageVersion('RAQSAPI')} library for R')
@@ -82,4 +78,4 @@ aqs_sign_up <- function(email)
   glue::glue("A verification email will be sent to {email}  \n") %>%
     message()
   return(invisible())
-}  #nocov end
+} #nocov end

[…]
