### 2026-05-21 7673d26
add NEWS.Rmd, add Ropensci badge to README, Rmarkdownchuns set to eval=FALSE in UsingRAQSAPI.Rmd, minor lint changes
