+++ NEWS.md
@@ -1,0 +1,17 @@
+---
+title: "News.md Please refer to crancomments.md"
+author: Clinton Mccrowey, physical scientist - US EPA  
+abstract: Please refer to cran-comments.[Rmd/md]
+output:
+  rmarkdown::github_document:
+    toc: true
+  rmarkdown::html_document: null
+lang: en-US
+bibliography: ./vignettes/AQSAPI.bib
+csl: ./vignettes/acs-nano.csl
+---
+
+
+
+
+Please refer to (cran-comments file)[cran-comments.md].
