### 2026-05-21 54a8815
fix trivial issues in *.Rmd files
