+++ vignettes/InstallandSetup.Rmd
@@ -36,6 +36,7 @@ pak::pkg_install(pkgs="RAQSAPI", dependencies = TRUE)
 ## Option 2: Installing the development version of RAQSAPI (unstable)
 
 > [!CAUTION]
+>
 > The development version of RAQSAPI contains code that has not yet been thoroughly been tested and is more
 > likely to contain bugs. It is recommended that the stable CRAN version be installed for production use.
 
