### 2026-05-26 4a67f3c
address YAML header issues in *.Rmd files
