+++ NEWS.Rmd
@@ -1,17 +1,17 @@
 ---
 title: "News.md Please refer to crancomments.md"
-author: Clinton Mccrowey, physical scientist - US EPA  
+author:
+- name: Clinton Mccrowey
+  affiliation: United States Environmental Protection Agency Region 3, Air and Radiation
 abstract: Please refer to cran-comments.[Rmd/md]
 output:
-  rmarkdown::github_document:
-    toc: true
-  rmarkdown::html_document: null
+  rmarkdown::github_document: null
+  rmarkdown::html_vignette: null
 lang: en-US
 bibliography: ./vignettes/AQSAPI.bib
 csl: ./vignettes/acs-nano.csl
 ---
 
-```{r setup, include=FALSE}
 ```{r SETUP, echo = FALSE, message = FALSE, warning = FALSE}
 knitr::opts_chunk$set(
   collapse = TRUE,
