+++ README.Rmd
@@ -1,22 +1,25 @@
 ---
 title: "Introduction to the RAQSAPI package"
-author: Clinton Mccrowey, physical scientist - US EPA  
-abstract: RAQSAPI is a package for R that connects the R programming language
-  environment to the United States Environmental Protection Agency's Air Quality
-  System (AQS) Data Mart database for retrieval of ambient air pollution data.
+author:
+- name: Clinton Mccrowey
+  affiliation: United States Environmental Protection Agency Region 3, Air and Radiation
+    Division
+abstract: RAQSAPI is a package for R that connects the R programming language environment
+  to the United States Environmental Protection Agency's Air Quality System (AQS)
+  Data Mart database for retrieval of ambient air pollution data.
 output:
   rmarkdown::github_document:
     toc: true
-  rmarkdown::html_document: null
-lang: en-US
+  rmarkdown::html_vignette: null
+lang: "en-US"
 bibliography: ./vignettes/AQSAPI.bib
-csl: ./vignettes/acs-nano.csl
+csl: "./vignettes/acs-nano.csl"
 ---
 ```{r SETUP, echo = FALSE, message = FALSE, warning = FALSE}
 knitr::opts_chunk$set(
   collapse = TRUE,
   comment = "#>",
-  tidy.ops = list(width.cutoff = 80),
+  tidy.ops = list(width.cutoff = 125),
   tidy = TRUE
 )
 ```
