+++ vignettes/InstallandSetup.Rmd
@@ -1,16 +1,12 @@
 ---
-title: RAQSAPI - Install and Setup
+title: "RAQSAPI - Install and Setup"
 author:
-- affiliation: |-
-    United States Environmental Protection Agency (US EPA),  
-        Region III,  
-        Air and Radiation Division,  
-        Air Quality and Analysis Branch
-  name: Clinton Mccrowey
+- name: Clinton Mccrowey
+  affiliation: United States Environmental Protection Agency Region 3, Air and Radiation
+    Division
 output:
-  md_document:
-    variant: gfm
-  rmarkdown::html_vignette: default
+  rmarkdown::github_document: null
+  rmarkdown::html_vignette: null
 vignette: >
   %\VignetteIndexEntry{RAQSAPI - Install and Setup}
   %\VignetteEncoding{UTF-8}
@@ -34,7 +30,7 @@ pak::pkg_install(pkgs="RAQSAPI", dependencies = TRUE)
 ```
 
 ## Option 2: Installing the development version of RAQSAPI (unstable)
-
+```{asis=true}
 > [!CAUTION]
 >
 > The development version of RAQSAPI contains code that has not yet been thoroughly been tested and is more
@@ -42,7 +38,7 @@ pak::pkg_install(pkgs="RAQSAPI", dependencies = TRUE)
 
 To install the development version of `RAQSAPI`, first if not already installed, install the `pak` package and its
 dependencies. Then run the following in a R environment.
-
+```
 ``` r
 pak::pkg_install("USEPA/RAQSAPI")
 ```
