+++ vignettes/RAQSAPIusagetipsandprecautions.Rmd
@@ -1,10 +1,12 @@
 ---
-title: RAQSAPI - Usage tips and precautions
-author: Clinton Mccrowey physical scientist - US EPA
+title: "RAQSAPI - Usage tips and precautions"
+author:
+- name: Clinton Mccrowey
+  affiliation: United States Environmental Protection Agency Region 3, Air and Radiation
+    Division
 output:
-  rmarkdown::html_vignette: default
-  md_document:
-    variant: gfm
+  rmarkdown::github_document: null
+  rmarkdown::html_vignette: null
 vignette: >
   %\VignetteIndexEntry{RAQSAPI - Usage tips and precautions}
   %\VignetteEncoding{UTF-8}
