+++ RAQSAPIvignette.md
@@ -1,119 +1,492 @@
 ---
-title: "RAQSAPI - RAQSAPI functions Long version"
+title: "Introduction to the RAQSAPI package"
 author:
-- name: Clinton Mccrowey
-  affiliation: United States Environmental Protection Agency Region 3, Air and Radiation
-    Division
+- affiliation: "United States Environmental Protection Agency (US EPA),  \n    Region
+    III,  \n    Air and Radiation Division,  \n    Air Quality and Analysis Branch"
+  name: Clinton Mccrowey
+abstract: RAQSAPI is a package for R that connects the R programming language environment
+  to the United States Environmental Protection Agency's (US EPA) Air Quality System
+  (AQS) Data Mart database API for retrieval of ambient air pollution data.
 output:
-  rmarkdown::github_document: null
-  rmarkdown::html_vignette: null
-vignette: >
-  %\VignetteIndexEntry{RAQSAPI - RAQSAPI functions Long version}
-  %\VignetteEncoding{UTF-8}
-  %\VignetteDepends{knitr}
-  %\VignetteDepends{rmarkdown}
-  %\VignetteEngine{knitr::rmarkdown}
+  html_document:
+    toc: true
+    df_print: paged
+  md_document:
+    variant: gfm
+  rmarkdown::html_vignette: default
+  pdf_document:
+    toc: true
+lang: "en-US"
+vignette: "%\\VignetteIndexEntry{Introduction to the RAQSAPI package} %\\VignetteEncoding{UTF-8}
+  %\\VignetteKeyword{R} %\\VignetteKeyword{AQS} %\\VignetteKeyword{EPA} %\\VignetteKeyword{Data
+  Mart} %\\VignetteKeyword{API} %\\VignetteDepends{magrittr} %\\VignetteDepends{stringr}
+  %\\VignetteDepends{tibble} %\\VignetteDepends{purrr} %\\VignetteDepends{knitr} %\\VignetteDepends{rmarkdown}
+  %\\VignetteDepends{RAQSAPI} %\\VignetteEngine{knitr::rmarkdown}\n"
+bibliography: AQSAPI.bib
+csl: "acs-nano.csl"
 ---
+
+
+<img src="./figures/RAQSAPIhexsticker.png" alt="RAQSAPI hexsticker" width="150" height="150">
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+# EPA Disclaimer
+
+> [!NOTE]
+>
+> This software/application was developed by the U.S. Environmental Protection Agency (USEPA). No warranty expressed or
+> implied is made regarding the accuracy or utility of the system, nor shall the act of distribution constitute any such
+> warranty. The USEPA has relinquished control of the information and no longer has responsibility to protect the integrity,
+> confidentiality or availability of the information. Any reference to specific commercial products, processes, or services
+> by service mark, trademark, manufacturer, or otherwise, does not constitute or imply their endorsement, recommendation or
+> favoring by the USEPA. The USEPA seal and logo shall not be used in any manner to imply endorsement of any commercial
+> product or activity by the USEPA or the United States Government.
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+>[!WARNING]
+>
+> US EPA's AQS Data Mart API V2 is currently
+> in beta phase of development, the API interface has not been finalized.
+> This means that certain functionality of the API may change or be removed
+> without notice. As a result, this package is also currently marked as beta and
+> may also change to reflect any changes made to the Data Mart API or in respect
+> to improvements in the design, functionality, quality and documentation of
+> this package. The authors assume no liability for any problems that may occur
[…]

[…]
