+++ vignettes/InstallandSetup.Rmd
@@ -30,7 +30,7 @@ pak::pkg_install(pkgs="RAQSAPI", dependencies = TRUE)
 ```
 
 ## Option 2: Installing the development version of RAQSAPI (unstable)
-```{asis=true}
+
 > [!CAUTION]
 >
 > The development version of RAQSAPI contains code that has not yet been thoroughly been tested and is more
@@ -38,7 +38,7 @@ pak::pkg_install(pkgs="RAQSAPI", dependencies = TRUE)
 
 To install the development version of `RAQSAPI`, first if not already installed, install the `pak` package and its
 dependencies. Then run the following in a R environment.
-```
+
 ``` r
 pak::pkg_install("USEPA/RAQSAPI")
 ```
