### 2026-05-26 61089d8
rebuild README.Rmd, addresses #36
