### 2026-05-26 b6c1d5a
update pkgcheck github workflaw to use actions/checkout version 6
