+++ .github/workflows/pkgcheck.yaml
@@ -21,7 +21,7 @@ jobs:
       issues: write
       pull-requests: read
     steps:
-      - uses: actions/checkout@v4
+      - uses: actions/checkout@v6
       - name: Find PR number
         id: pr
         run: |
