### 2026-07-09 e96994f
Remove gtools dependency, Fix broken reference in CODE_OF_CONDUCT.md, update DESCRIPTION with reverer information, fix invalid structure in codemeta.json, minor typo in checkaqsparams function, update inst/WORDLIST
