+++ vignettes/RAQSAPIPurpose.Rmd
@@ -60,62 +60,6 @@ AQS Data Mart once per week and this data is made available to the public
 through web-based applications and APIs (application programming interface)
 [@AQSDataMartWelcome]. RAQSAPI functions use the APIs provided by AQS Data Mart
 to retrieve data.
-
-<!-- # The R programming language -->
-<!--  The R language and it's environment [@website:whatisR] (Hereafter referred to -->
-<!--  as simply R) is a domain-specific programming environment that is inspired by -->
-<!--  the S statistical language and it's respective environment which was developed -->
-<!--  at Bell Laboratories (formerly AT&T, now Lucent Technologies) by John Chanmbers -->
-<!--  et al. The first public release of the S programming environment was released -->
-<!--  in 1976 [@website:whatisR]. -->
-
-<!--  The reference implementation of R has been accepted as a GNU project -->
-<!--  [@website:GNUproject] which means the code is free//libre open source software -->
-<!--  (FLOSS) under the terms of the [GNU General Public -->
-<!--  License](https://www.r-project.org/COPYING) (GNU GPL) [@website:whatisR]. -->
-
-<!-- ## Why R? -->
-<!--  The R environment is available for most of the major computer operating -->
-<!--  systems - Windows, unix and \*nix-like platforms such as Solaris, GNU\\Linux, -->
-<!--  \*BSD, Darwin-derived platforms such as Mac OS X and other POSIX and POSIX-like -->
-<!--  operating systems. -->
-
-<!--  R also is an well-established statistical programming environment that has wide -->
-<!--  adoption in the statistical, academic, and data analytics communities because -->
-<!--  of it's good third party support, a large user community that accomodates users -->
-<!--  from all experience levels and multilanguage support. -->
-
-<!--  R is easily extensible, in fact there is a large collection of Free packages -->
-<!--  available that extends the capabilities of the base R environment. The largest -->
-<!--  collection of packages available for R is hosted by -->
-<!--  [The Comprehensive R Archive Network (CRAN)](https://cran.r-project.org), -->
-<!--  currently CRAN hosts over 15,000 packages and are available to anyone under -->
-<!--  permissive licensing terms. These packages are developed and maintained by the -->
-<!--  worldwide R user community. -->
-
-<!-- * The design of the R statistical programming environment makes it a great -->
-<!--     platform for statistical calculation, data manipulation and graphical -->
-<!--     display of data. The R includes: -->
-<!--   + Packages such as feather, foreign, dbplyr, readR, which can be used to -->
-<!--       import data from external sources. -->
-<!--   + R is has support from many different integrated development environments to -->
-<!--       ease development including Visual Studio Code, Rstudio, Eclipse, Rattle -->
-<!--       GUI, R Commander, RKWard, Jupyter, StatET. R syntax highlighting is also -->
-<!--       supported in many common text and source code editors either natively or -->
-<!--       via an application add-on such as Emacs, Vim, Notepad++, Sublime Text -->
-<!--       and Atom. -->
-<!--   + R has a large set of operators for performing calculations on arrays and -->
-<!--       tabular data sets. -->
-<!--       a.  A functional programming style that makes it easy to apply a single -->
-<!--       operation on large data sets. -->
-<!--   + Tools for graphical output of data such as knitr and rmarkdown for creating -->
-<!--   reports, esquisse and ggplot2 creating graphs, openair for creating windroses. -->
-<!--   Rshiny for a web interface. -->
-<!--   + Linear and non-linear modeling tools. -->
-<!--   + R has hooks into many popular software packages used for geospatial analysis. -->
-
-<!--   These tools make R an excellent environment to perform statistical analysis on -->
-<!--   ambient air monitoring data. -->
   
 # Benefits of using RAQSAPI
 RAQSAPI package for R is an easy to use R package that allows
