+++ CODE_OF_CONDUCT.md
@@ -57,7 +57,8 @@ representative at an online or offline event.
 ## Enforcement
 
 Instances of abusive, harassing, or otherwise unacceptable behavior may be
-reported to the community leaders responsible for enforcement at {{{ contact }}}.
+reported to the on 
+[RAQSAPI's github discussion board on github](https://github.com/USEPA/RAQSAPI/discussions).
 All complaints will be reviewed and investigated promptly and fairly.
 
 All community leaders are obligated to respect the privacy and security of the
