+++ NAMESPACE
@@ -95,7 +95,6 @@ importFrom(dplyr,rename_at)
 importFrom(dplyr,select)
 importFrom(dplyr,vars)
 importFrom(glue,glue)
-importFrom(gtools,invalid)
 importFrom(httr2,last_response)
 importFrom(httr2,req_error)
 importFrom(httr2,req_options)
@@ -133,6 +132,7 @@ importFrom(rlang,call_name)
 importFrom(rlang,caller_call)
 importFrom(rlang,format_error_bullets)
 importFrom(rlang,is_character)
+importFrom(rlang,is_empty)
 importFrom(stringr,str_c)
 importFrom(stringr,str_detect)
 importFrom(stringr,str_replace_all)

[…]
