### 2026-07-13 83f8e5a
remove vignette entries from vigeteete sections, consolidate the number of vignettes to 2, move vignettes setions to vignettes/includes/
