+++ tests/spelling.R
@@ -1,3 +1,6 @@
 if (requireNamespace("spelling", quietly = TRUE)) {
-  spelling::spell_check_test(vignettes = FALSE, error = FALSE, lang = "en-US", skip_on_cran = FALSE)
+  spelling::spell_check_test(vignettes = TRUE,
+                             error = FALSE,
+                             lang = "en-US",
+                             skip_on_cran = TRUE)
 }
