+++ vignettes/includes/RAQSAPIFunctions-Brief.Rmd
@@ -6,13 +6,7 @@ author:
     Division
 output:
   rmarkdown::github_document: null
-  rmarkdown::html_vignette: null
-vignette: >
-  %\VignetteIndexEntry{RAQSAPI - A Brief introduction into RAQSAPI functions}
-  %\VignetteEncoding{UTF-8}
-  %\VignetteDepends{knitr}
-  %\VignetteDepends{rmarkdown}
-  %\VignetteEngine{knitr::rmarkdown}
+  rmarkdown::html_document: null
 ---
 
 RAQSAPI functions are named according to the service and filter variables that
