+++ README.Rmd
@@ -12,8 +12,8 @@ output:
     toc: true
   rmarkdown::html_vignette: null
 lang: "en-US"
-bibliography: ./vignettes/AQSAPI.bib
-csl: "./vignettes/acs-nano.csl"
+bibliography: ./vignettes/includes/AQSAPI.bib
+csl: "./vignettes/includes/acs-nano.csl"
 ---
 ```{r SETUP, echo = FALSE, message = FALSE, warning = FALSE}
 knitr::opts_chunk$set(
@@ -49,28 +49,28 @@ https://choosealicense.com/licenses/mit/)
 https://github.com/ropensci/software-review/issues/744)
 <!-- badges: end -->
 
-```{r child = "./vignettes/EPA_Disclaimer.Rmd", echo = FALSE, comment = NA}
+```{r child = "./vignettes/includes/EPA_Disclaimer.Rmd", echo = FALSE, comment = NA}
 ```
 
-```{r child = "./vignettes/Intro.Rmd"}
+```{r child = "./vignettes/includes/Intro.Rmd"}
 ```
 
-```{r child = "./vignettes/TimelinessofAQSData.Rmd"}
+```{r child = "./vignettes/includes/TimelinessofAQSData.Rmd"}
 ```
 
-```{r child = "./vignettes/InstallandSetup.Rmd", echo = FALSE, comment = NA}
+```{r child = "./vignettes/includes/InstallandSetup.Rmd", echo = FALSE, comment = NA}
 ```
 
-```{r child = "./vignettes/UsingRAQSAPI.Rmd"}
+```{r child = "./vignettes/includes/UsingRAQSAPI.Rmd"}
 ```
 
-```{r child = "./vignettes/RAQSAPIFunctions-Brief.Rmd"}
+```{r child = "./vignettes/includes/RAQSAPIFunctions-Brief.Rmd"}
 ```
 
-```{r child = "./vignettes/pyaqsapi.Rmd"}
+```{r child = "./vignettes/includes/pyaqsapi.Rmd"}
 ```
 
-```{r child = "./vignettes/Acknowledgements.Rmd"}
+```{r child = "./vignettes/includes/Acknowledgements.Rmd"}
 ```
 
 # References

[…]
