### 2026-07-14 1254e2d
created a todo list of project improvements as suggested by RopenSci
