+++ tests/testthat.R
@@ -1,4 +1,5 @@
 library(testthat)
 library(RAQSAPI)
+library(checkmate)
 
 test_check("RAQSAPI")
