### 2026-07-14 cdfdd7e
update RopenSci todo list
