+++ Ropensci Review of RAQSAPI - todo list.md
@@ -3,37 +3,37 @@
 ## Must fix
 
 | Status | Item | Why it matters | Source(s) | submitter comments |
-|---|---|---|---|---|
+|:---:|:---:|:---:|:---:|:---:|
 |  | Fix `checkaqsparams()` so validation actually runs for the production wrapper calls | Right now positional calls appear to skip validation, so invalid inputs may pass silently | Milan Malfait (@milanmlft); Rainer M Krug (@rkrug) |  |
-|&check;|| Fix the always-true MA/pqao validation logic in `checkaqsparams()` | The current logical condition would reject valid values if reached | Milan Malfait (@milanmlft) |  |
+|&check;| Fix the always-true MA/pqao validation logic in `checkaqsparams()` | The current logical condition would reject valid values if reached | Milan Malfait (@milanmlft) |  |
 |  | Fix the broken `rlang::abort()` path in validation | The validator appears to fail internally instead of showing the intended error message | Milan Malfait (@milanmlft); Rainer M Krug (@rkrug) |  |
 |  | Fix the S3 validator bug in `AQS_DATAMART_APIv2` | The second data slot is not being validated correctly | Milan Malfait (@milanmlft) |  |
 |  | Remove the `Sys.setenv()` mutation from `.onLoad()` and clean up the ineffective save/restore logic in `.onUnLoad()` | Packages should not mutate user session-wide environment settings on load | Milan Malfait (@milanmlft) |  |
 |  | Replace the empty / ineffective `NEWS.md` with a real changelog | `R CMD check` reports no news entries and the file is effectively a stub | Milan Malfait (@milanmlft); Rainer M Krug (@rkrug) |  |
 |  | Strengthen tests beyond `expect_no_error()` smoke tests | The current suite can pass even if returned data are wrong | Milan Malfait (@milanmlft); Rainer M Krug (@rkrug) |  |
-|  | Add tests for validation failure paths, especially `checkaqsparams()` | This would catch the validation bugs and protect future fixes | Milan Malfait (@milanmlft) |  |
+|  | Add tests for validation failure paths, especially `checkaqsparams()` | This would catch the validation bugs and protect future fixes | Milan Malfait (@milanmlft) | Please refer to [tests/testthat/test-helperfunctions.R](https://github.com/USEPA/RAQSAPI/blob/main/tests/testthat/test-helperfunctions.R#L25-L66) |
 |  | Fix Rd rendering issues such as leaked `@seealso`, broken math markup, and the misleading `aqs_sign_up` note | These are user-facing documentation defects | Milan Malfait (@milanmlft) |  |
 |  | Restore or add a proper `CONTRIBUTING` file | Community guidelines are incomplete without it | Milan Malfait (@milanmlft); Rainer M Krug (@rkrug) |  |
 |  | Fix malformed package metadata and validation blockers in `DESCRIPTION` | A bad `DESCRIPTION` comment/version field caused `devtools::check()` to fail; metadata hygiene needs cleanup | Milan Malfait (@milanmlft); Rainer M Krug (@rkrug) |  |
-|  | Regenerate or fix `codemeta.json` | The file is invalid JSON as reviewed | Milan Malfait (@milanmlft) |  |
-|  | Fix `CODE_OF_CONDUCT.md` placeholders and broken links | Community files currently contain template leftovers and malformed markdown | Milan Malfait (@milanmlft) |  |
-|  | Delete stale `tests/test-RAQSAPI.R.blob` | It is obsolete and confusing, even if it does not break checks | Milan Malfait (@milanmlft) |  |
[…]
